[Notification] refactor: History, Messaging 분리 #114
