+++ notification/src/main/java/com/bobjool/notification/application/service/HistoryService.java
@@ -1,7 +1,6 @@
 package com.bobjool.notification.application.service;
 
 
-import com.bobjool.notification.application.dto.NotificationDto;
 import com.bobjool.notification.application.dto.NotificationHistoryDto;
 import com.bobjool.notification.application.dto.NotificationSearchDto;
 import com.bobjool.notification.domain.entity.Notification;
@@ -17,13 +16,11 @@ import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
-public class NotificationService {
-    private final SlackService slackService;
+public class HistoryService {
     private final NotificationRepository notificationRepository;
     private final TemplateService templateService;
 
     public Page<NotificationHistoryDto> searchNotification(NotificationSearchDto searchDto, Pageable pageable) {
-
         Page<Notification> notificationPage = notificationRepository.search(
                 searchDto.userId(),
                 searchDto.message(),
@@ -39,37 +36,12 @@ public class NotificationService {
         );
     }
 
-    @Transactional
-    public void postNotification(NotificationDto dto) {
-        switch (dto.messageChannel()) {
-            case SLACK:
-                sendDirectMessage(
-                        dto.userSlack(),
-                        dto.messageTitle().concat(
-                                dto.messageContent())
-                );
-                break;
-            case EMAIL:
-            default:
-                sendEmail(dto.userEmail(), dto.messageTitle(), dto.messageContent());
-        }
-
-    }
 
     @Transactional
     protected void saveNotification(UUID templateId, Long userId, String jsonData, String message, String userContact) {
         Template template = templateService.selectTemplateEntity(templateId);
         Notification notification = Notification.createNotification(template, userId, jsonData, message, userContact);
         notificationRepository.save(notification);
-    }
-
-    private void sendDirectMessage(String slackId, String message) {
-        String channelId = slackService.conversationsOpen(slackId);
-        slackService.chatPostMessage(channelId, message);
-    }
-
-    private void sendEmail(String email, String title, String content) {
-        // TODO. Email 라이브러리 적용
     }
 
 }
